fix jl_fs_write type sig
uv_fs_write:offset is int64_t, not size_t.

Closes nolta/Winston.jl#182.

+++ src/jl_uv.c
@@ -452,7 +452,7 @@ DLLEXPORT int jl_fs_chmod(char *path, int mode)
     return ret;
 }
 
-DLLEXPORT int jl_fs_write(int handle, char *data, size_t len, size_t offset)
+DLLEXPORT int jl_fs_write(int handle, char *data, size_t len, int64_t offset)
 {
     uv_fs_t req;
     uv_buf_t buf[1];
